// product/frontend/DNAustria/src/app/api/model/createContactDto.ts
/**
 * DNAustria.Api | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateContactDto { 
    name: string;
    email?: string | null;
    phone?: string | null;
    organization: string | null;
}